# api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Account is suspended or banned"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Account attached to the current token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/admin/profile/me": {
      "get": {
        "operationId": "AdminProfileController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not an admin, or IP not whitelisted"
          },
          "404": {
            "description": "Account has no profile yet"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Your own admin profile",
        "tags": [
          "Admin-Profile"
        ]
      },
      "patch": {
        "operationId": "AdminProfileController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not an admin, or IP not whitelisted"
          },
          "404": {
            "description": "Account has no profile yet"
          },
          "409": {
            "description": "Email already taken"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Partially update your own admin profile",
        "tags": [
          "Admin-Profile"
        ]
      }
    },
    "/admin/profile/me/change-password": {
      "patch": {
        "operationId": "AdminProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "New password is weak, or same as the current one"
          },
          "401": {
            "description": "Token invalid, or current password is incorrect"
          },
          "403": {
            "description": "Not an admin, or IP not whitelisted"
          },
          "404": {
            "description": "Account not found"
          },
          "429": {
            "description": "Too many attempts"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change your own password. Revokes every OTHER session, keeps this one",
        "tags": [
          "Admin-Profile"
        ]
      }
    },
...
    "/brand/profile/me": {
      "get": {
        "operationId": "BrandProfileController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not a brand account"
          },
          "404": {
            "description": "Account has no profile yet"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Your own brand profile",
        "tags": [
          "Brand"
        ]
      },
      "patch": {
        "operationId": "BrandProfileController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not a brand account"
          },
          "404": {
            "description": "Account has no profile yet"
          },
          "409": {
            "description": "Email or tax code already taken"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Partially update your own brand profile",
        "tags": [
          "Brand"
        ]
      },
      "delete": {
        "operationId": "BrandProfileController_deleteMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not a brand account"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account still has collaborations"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete your own brand account. Irreversible",
        "tags": [
          "Brand"
        ]
      }
    },
    "/brand/profile/me/change-password": {
      "patch": {
        "operationId": "BrandProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "New password is weak, or same as the current one"
          },
          "401": {
            "description": "Token invalid, or current password is incorrect"
          },
          "403": {
            "description": "Not a brand account"
          },
          "404": {
            "description": "Account not found"
          },
          "429": {
            "description": "Too many attempts"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change your own password. Revokes every OTHER session, keeps this one",
        "tags": [
          "Brand"
        ]
      }
    },
    "/creator/profile/me": {
      "get": {
        "operationId": "CreatorProfileController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not a creator account"
          },
          "404": {
            "description": "Account has no profile yet"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Your own creator profile",
        "tags": [
          "Creator-Profile"
        ]
      },
      "patch": {
        "operationId": "CreatorProfileController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCreatorProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, invalid or expired"
          },
          "403": {
            "description": "Not a creator account"
          },
          "404": {
            "description": "Account has no profile yet"
          },
          "409": {
            "description": "Email already taken"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Partially update your own creator profile",
        "tags": [
          "Creator-Profile"
        ]
      }
    },
    "/creator/profile/me/change-password": {
      "patch": {
        "operationId": "CreatorProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "New password is weak, or same as the current one"
          },
          "401": {
            "description": "Token invalid, or current password is incorrect"
          },
          "403": {
            "description": "Not a creator account"
          },
          "404": {
            "description": "Account not found"
          },
          "429": {
            "description": "Too many attempts"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change your own password. Revokes every OTHER session, keeps this one",
        "tags": [
          "Creator-Profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "CRM KOC API",
    "description": "REST API for the CRM KOC system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ERole": {
        "type": "string",
        "enum": [
          "admin",
          "brand",
          "creator"
        ]
      },
      "EAccountStatus": {
        "type": "number",
        "enum": [
          1,
          2,
          3,
          4
        ]
      },
...
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "accountRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ERole"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EAccountStatus"
              }
            ]
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "emailVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "phoneVerifiedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "statusReason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "accountRole",
          "status",
          "phone",
          "emailVerifiedAt",
          "phoneVerifiedAt",
          "statusReason",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AdminProfileResponseDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "description": "Same value as accounts.id"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Ho_Chi_Minh"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "name",
          "email",
          "avatarUrl",
          "timezone",
          "updatedAt"
        ]
      },
      "UpdateAdminProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": "John Doe",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "example": "admin@gmail.com",
            "format": "email"
          },
          "avatarUrl": {
            "type": "object",
            "example": "https://cdn.example.com/a.png"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Ho_Chi_Minh",
            "maxLength": 64
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "description": "Mật khẩu hiện tại, dùng để xác thực lại"
          },
          "newPassword": {
            "type": "string",
            "example": "Abc@12345"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
...
      "BrandProfileResponseDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "description": "Same value as accounts.id"
          },
          "brandName": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "companyName": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "taxCode": {
            "type": "string",
            "nullable": true,
            "maxLength": 14
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 20
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "industry": {
            "type": "string",
            "nullable": true,
            "maxLength": 64
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "contactPhone": {
            "type": "string",
            "nullable": true,
            "maxLength": 20
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Ho_Chi_Minh"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "brandName",
          "companyName",
          "taxCode",
          "email",
          "phone",
          "website",
          "industry",
          "logoUrl",
          "description",
          "address",
          "contactName",
          "contactPhone",
          "timezone",
          "updatedAt"
        ]
      },
      "UpdateBrandProfileDto": {
        "type": "object",
        "properties": {
          "brandName": {
            "type": "object",
            "example": "Cocoon",
            "maxLength": 255
          },
          "companyName": {
            "type": "object",
            "example": "Cong ty TNHH My pham Cocoon Viet Nam",
            "maxLength": 255
          },
          "taxCode": {
            "type": "object",
            "example": "0123456789",
            "maxLength": 14
          },
          "email": {
            "type": "string",
            "example": "contact@cocoon.vn",
            "format": "email"
          },
          "phone": {
            "type": "object",
            "example": "+84900000001",
            "maxLength": 20
          },
          "website": {
            "type": "object",
            "example": "https://cocoon.vn"
          },
          "industry": {
            "type": "object",
            "example": "beauty",
            "maxLength": 64
          },
          "logoUrl": {
            "type": "object",
            "example": "https://cdn.example.com/logo.png"
          },
          "description": {
            "type": "object",
            "example": "Thuong hieu my pham thuan chay"
          },
          "address": {
            "type": "object",
            "example": "12 Nguyen Hue St, District 1, Ho Chi Minh City"
          },
          "contactName": {
            "type": "object",
            "example": "Nguyen Van A",
            "maxLength": 255
          },
          "contactPhone": {
            "type": "object",
            "example": "+84900000002",
            "maxLength": 20
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Ho_Chi_Minh",
            "maxLength": 64
          }
        }
      },
      "ECreatorContent": {
        "type": "string",
        "enum": [
          "beauty",
          "personal_care",
          "fashion",
          "food",
          "travel",
          "experience_culture",
          "mom_baby",
          "family",
          "lifestyle_home",
          "health",
          "education",
          "life_skills",
          "finance",
          "entertainment",
          "art",
          "gaming",
          "streamer",
          "business",
          "b2b",
          "sales",
          "fitness",
          "pet",
          "automotive",
          "music"
        ]
      },
      "CreatorProfileResponseDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "description": "Same value as accounts.id"
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 20
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "gender": {
            "type": "number",
            "nullable": true,
            "description": "1 = male, 2 = female, 3 = other"
          },
          "city": {
            "type": "string",
            "nullable": true,
            "maxLength": 128
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "contentCategories": {
            "type": "array",
            "example": [
              "beauty",
              "fashion"
            ],
            "items": {
              "$ref": "#/components/schemas/ECreatorContent"
            }
          },
          "portfolioUrl": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Ho_Chi_Minh"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "displayName",
          "email",
          "phone",
          "bio",
          "avatarUrl",
          "dateOfBirth",
          "gender",
          "city",
          "address",
          "contentCategories",
          "portfolioUrl",
          "timezone",
          "updatedAt"
        ]
      },
      "UpdateCreatorProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "object",
            "example": "Chau Bui",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "example": "creator@gmail.com",
            "format": "email"
          },
          "phone": {
            "type": "object",
            "example": "+84900000003",
            "maxLength": 20
          },
          "bio": {
            "type": "object",
            "example": "KOC linh vuc lam dep va thoi trang"
          },
          "avatarUrl": {
            "type": "object",
            "example": "https://cdn.example.com/a.png"
          },
          "dateOfBirth": {
            "type": "object",
            "example": "1998-04-21",
            "format": "date"
          },
          "gender": {
            "type": "number",
            "example": 1,
            "description": "1 = male, 2 = female, 3 = other"
          },
          "city": {
            "type": "object",
            "example": "Ho Chi Minh City",
            "maxLength": 128
          },
          "address": {
            "type": "object",
            "example": "12 Nguyen Hue St, District 1, Ho Chi Minh City"
          },
          "contentCategories": {
            "type": "array",
            "example": [
              "beauty",
              "fashion"
            ],
            "maxItems": 20,
            "items": {
              "$ref": "#/components/schemas/ECreatorContent"
            }
          },
          "portfolioUrl": {
            "type": "object",
            "example": "https://portfolio.example.com"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Ho_Chi_Minh",
            "maxLength": 64
          }
        }
      },
...
    }
  }
}